COMMERCE-4881 dispatch-api - Rename
